+++ gon/oriented.py
@@ -1,0 +1,13 @@
+from abc import abstractmethod
+
+from .angular import Orientation
+from .geometry import Geometry
+
+
+class Oriented(Geometry):
+    @property
+    @abstractmethod
+    def orientation(self) -> Orientation:
+        """
+        Returns orientation of the geometry.
+        """
